removed excluding if either replicate is below 0.2
I don't think that makes sense in the context of our work like it does with the ELISA data. We are about to set what values are seropositive so I think that takes care of it. If we do want an exclusion minimum value for each replicate individually than I don't think it should necessarily be log2(0.2)...., maybe zero? Since then one of the replicates would have less signal than the buffer?