dealing with duplicates in sample metadata (epi data)
The duplicates which are just repeated rows (exact same info 2x), one of them is removed and the other kept. For duplicates which are not matching epi info, they are called out, exported as a table, automatically removed from the metadata. Those duplicates are set to exclude = yes in the sample data frame so that they can be removed from analysis later. Then the sample data frame and the metadata frame are merged. All duplicates in the assay are kept.