need to change how the annotation targets data frame is created to make it more universal
Because I sorted the sanger sample list by sample_id rather than having slide 1 block 1 listed 1st, the numbers attached to the targets were coming out all weird because it called up a sample in the middle of the slides_all.df. This will need to be changed in the primary (master) script as well.