modified to include setting negative normalized values to zero
Now, we have both setting negative normalized values to Zero and not doing that. The negative values are included in exported data to send to Nuno.

The negative values are set to 0 for analysis in the data analysis section at the bottom of the script.

The replacing target names thing at the bottom isn't working yet.  R just crashed so I am committing what I have changed so far.