added saving "slides_all.df" as an R object
It saves time to load this obejct rather than re-run the part of the script that matches the sample to the slides_all database.

I think we are at the point where we might want to reorganize the script in sections, where you save and load objects for different sections to avoid running the whole thing every time.