Added setting the negative normalized values to 0 in main script
It is now structured so that we carry forward negative normalized values and also set to zero in different parts of the later script. So we have to average duplicates separately in both cases. Negative normalized values are used for the std 3 plot and also to send to Nuno.  The 0 and positive values only are used in our "data analysis" section.