Switched order of removing control samples and targets and excluded samples during analysis
The order for removing control targets and samples was after removing excluded samples. Because the samples_test vector is a logic vector, the numbers it called out would have been altered by removing the excluded samples, so needed to do that first. Also deleted samples_sub because we didn't use that for anything later.